rename action-recommendations route to activation-recommendations
