+++ front-api/routes/w/[wId]/index.ts
@@ -27,7 +27,7 @@ import { validate } from "@front-api/middlewares/validator";
 import { workspaceAuth } from "@front-api/middlewares/workspace_auth";
 import { escape } from "html-escaper";
 import { z } from "zod";
-import actionRecommendations from "./action-recommendations";
+import actionRecommendations from "./activation-recommendations";
 import analytics from "./analytics";
 import assistant from "./assistant";
 import auditLogs from "./audit-logs";
@@ -952,7 +952,7 @@ app.post(
 
 // Sub-apps using the catch-all default + the partial-subtree exception
 // targets declared above.
-app.route("/action-recommendations", actionRecommendations);
+app.route("/activation-recommendations", actionRecommendations);
 app.route("/analytics", analytics);
 app.route("/model_tiers", modelTiers);
 app.route("/assistant", assistant);
